+++ testcafe/page-object/common/breadcrumbs.ts
@@ -1,0 +1,9 @@
+import { t } from 'testcafe';
+
+export class Breadcrumbs {
+    constructor(public element: Selector) {}
+
+    public async goTo(name: string) {
+        await t.click(this.element.find('a').withText(name));
+    }
+}