+++ testcafe/page-object/editor/node-browser/node-browser-row.ts
@@ -1,0 +1,12 @@
+import { t } from 'testcafe';
+
+export class NodeBrowserRow {
+    public checkBox: Selector;
+    constructor(public element: Selector) {
+        this.checkBox = element.find('gtx-checkbox input');
+    }
+
+    public async open() {
+        await t.click(this.element.find('.displayName a'));
+    }
+}